fix escaping of xml special chars in query values
resolves an issue where I could not query for:
something from: <email@emailaddress.com>: the subject

The trick is to escape xml twice.  It's already escaped once outbound by
the suds parser, so we just escape all values as they come into the
query constructor.

It's possible the print output should be fixed to reflect the input
rather than what is submitted.